Renamed Controller/Service and Handled review comments
